:recycle: refactor(`<App>`): Extract `<Router>`.
This allows to avoid importing `<App>` in test fixtures.
